requirements.txt
# local package
-e .

# external requirements
Click==7.0
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
pandas==0.25.2
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scipy==1.3.1
seaborn==0.9.0
six==1.12.0
tqdm==4.36.1
urllib3==1.25.6
